chore: rename doc example
