+++ editor/example/doc-examples/overlay-tracks-bar-point.ts
