+++ editor/example/doc-examples/overlay-tracks-rect-text.ts
